EDU-204 Bonus change: change user controller in new controller format
